List, Detail에 작성자이름, include 사용법

+++ src/main/webapp/jsp/article/detail.jsp
@@ -3,11 +3,11 @@
 
 <%@ page language="java" contentType="text/html; charset=UTF-8"
     pageEncoding="UTF-8"%>
-
+    
 <%
 	Map<String, Object> articleMap = (Map<String, Object>) request.getAttribute("articleMap");
+	int loginedMemberId = (int) request.getAttribute("loginedMemberId");
 %>
-    
 <!DOCTYPE html>
 <html>
 <head>
@@ -17,15 +17,19 @@
 <body>
 	<h1><%= articleMap.get("id") %>번 게시물 상세보기</h1>
 	
-	<div> 번호 : <%= (int) articleMap.get("id") %></div>
-	<div> 작성일 : <%= (LocalDateTime) articleMap.get("regDate") %></div>
-	<div> 수정일 : <%= (LocalDateTime) articleMap.get("updateDate") %></div>
-	<div> 제목 : <%= (String) articleMap.get("title") %></div>
-	<div> 내용 : <%= (String) articleMap.get("body") %></div>
+	<div>번호 : <%= (int) articleMap.get("id") %></div>
+	<div>작성일 : <%= (LocalDateTime) articleMap.get("regDate") %></div>
+	<div>수정일 : <%= (LocalDateTime) articleMap.get("updateDate") %></div>
+	<div>작성자 : <%= (String) articleMap.get("writerName") %></div>
+	<div>제목 : <%= (String) articleMap.get("title") %></div>
+	<div>내용 : <%= (String) articleMap.get("body") %></div>
+	
 	<div>
 		<a href="list">목록</a>
-		<a href="modify?id=<%= (int) articleMap.get("id") %>">수정</a>
-		<a href="delete?id=<%= (int) articleMap.get("id") %>" onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false;">삭제</a>
+		<% if (loginedMemberId != -1 && loginedMemberId == (int) articleMap.get("memberId")) { %>
+			<a href="modify?id=<%= (int) articleMap.get("id") %>">수정</a>
+			<a href="delete?id=<%= (int) articleMap.get("id") %>" onclick="if(confirm('정말 삭제하시겠습니까?') == false) return false;">삭제</a>
+		<% } %>
 	</div>
 </body>
 </html>
